refactor: implement controller architecture and update local environment docs
